+++ Fossil/src/GameObject.cpp
@@ -1,42 +1,44 @@
 #include "GameObject.h"
 
-GameObject::GameObject() {}
+GameObject::GameObject() { logger.logWarn("GameObject.cpp: GameObject | No Args Constructor"); }
 
 GameObject::GameObject(std::string name, std::string file)
 {
+	logger.logWarn("GameObject.cpp: GameObject | Constructor");
 	ObjectName = name;
-	logger.logWarn(ObjectName + " Constructed\n");
 	ImportOBJ();
 }
 
-GameObject::~GameObject() {}
+GameObject::~GameObject() { logger.logWarn("GameObject.cpp: GameObject | Constructor"); }
 
 //	Import vertices from a 3d model
 void GameObject::ImportOBJ()
 {
-	vertices.at(0) = -0.5;
-	vertices.at(1) = -0.5;
-	vertices.at(2) = 0.0;
-	vertices.at(3) = 0.0;
-	vertices.at(4) = 0.5;
-	vertices.at(5) = 0.0;
-	vertices.at(6) = 0.5;
-	vertices.at(7) = -0.5;
-	vertices.at(8) = 0.0;
-	textures.at(0) = 255;
-	textures.at(1) = 0.0;
-	textures.at(2) = 0.0;
-	textures.at(3) = 0.0;
-	textures.at(4) = 255;
-	textures.at(5) = 0.0;
-	textures.at(6) = 0.0;
-	textures.at(7) = 0.0;
-	textures.at(8) = 255;
-	logger.logWarn("OBJ Imported\n");
+	logger.logWarn("GameObject.cpp: GameObject | ImportOBJ");
+	vertices.push_back(-0.5);
+	vertices.push_back(-0.5);
+	vertices.push_back(0.0);
+	vertices.push_back(0.0);
+	vertices.push_back(0.5);
+	vertices.push_back(0.0);
+	vertices.push_back(0.5);
+	vertices.push_back(-0.5);
+	vertices.push_back(0.0);
+	textures.push_back(255);
+	textures.push_back(0.0);
+	textures.push_back(0.0);
+	textures.push_back(0.0);
+	textures.push_back(255);
+	textures.push_back(0.0);
+	textures.push_back(0.0);
+	textures.push_back(0.0);
+	textures.push_back(255);
 }
 
 void GameObject::ConvertToArrays(GLfloat *verts, GLfloat *txtrs, GLfloat *norms, int *face)
 {
+	logger.logWarn("GameObject.cpp: GameObject | ConvertToArrays");
+
 	//	Copy vertices vector
 	for (int i = 0; i < vertices.size(); i++) {
 		verts[i] = vertices.at(i);