+++ Fossil/src/GameState.cpp
@@ -1,8 +1,6 @@
 #include "GameState.h"
 
-GameState::GameState()
-{
-}
+GameState::GameState() { logger.logWarn("GameState.cpp: GameState | No Args Constructor"); }
 
 GameState::~GameState()
 {