+++ Fossil/src/GameState.h
@@ -1,19 +1,19 @@
 #pragma once
 
+#include <vector>
 #include <string>
 #include "BuildDll.h"
 #include "GameObject.h"
-#include <vector>
 #include "glew.h"
 #include "glfw3.h"
 
 class FSL_API GameState
 {
 public:
+	std::string StateName;
 	std::vector<GameObject*> Objects;
-
 	GLFWwindow* Window;
-
+	
 	virtual ~GameState();
 
 	virtual void InitializeState() = 0;
@@ -23,8 +23,17 @@ public:
 
 	virtual void UpdateState() = 0;
 	virtual void DisplayState() = 0;
+	/*
+	~GameState();
 
+	void InitializeState();
+
+	void PauseState();
+	void ResumeState();
+
+	void UpdateState();
+	void DisplayState();
+	*/
 protected:
-	std::string StateName;
 	GameState();
 };