+++ Workspace/src/Project.cpp
@@ -6,19 +6,21 @@ class GS_StartScreen : public GameState
 {
 public:
 
-	GS_StartScreen() : GameState(){}
+	GS_StartScreen() : GameState() { wkspcLog.logWarn("Project.cpp: GS_StartScreen | No Args Constructor"); }
 
 	GS_StartScreen(std::string name, GLFWwindow* newWindow)
 	{
-		wkspcLog.logWarn("Start Screen Created\n");
+		wkspcLog.logWarn("Project.cpp: GS_StartScreen | Constructor");
 
 		StateName = name;
 		Window = newWindow;
 
-		Objects.at(0) = new GameObject("Triangle", "NoFileYet");
+		Objects.push_back(new GameObject("Triangle", "NoFileYet"));
 	}
 
-	void InitializeState() {}
+	~GS_StartScreen() { wkspcLog.logWarn("Project.cpp: GS_StartScreen | Deconstructor"); }
+
+	void InitializeState() { wkspcLog.logWarn("Project.cpp: GS_StartScreen | InitializeState"); }
 
 	void PauseState() {}
 
@@ -36,16 +38,18 @@ public:
 	{
 		Name = "Sub to Pewds";
 
-		wkspcLog.logWarn("Workspace Created\n");
+		wkspcLog.logWarn("Project.cpp: Workspace | No Args Constructor");
 
 		//	Push start up screen to the stack
 		PushState(new GS_StartScreen("Start Screen", GameWindow));
+		//States.push_back(new GS_StartScreen("Start Screen", GameWindow));
 	}
 
-	~Workspace() {}
+	~Workspace() { wkspcLog.logWarn("Project.cpp: Workspace | Deconstructor"); }
 };
 
 //	Create new game pointer
 Game* CreateGame() {
+	wkspcLog.logWarn("Project.cpp: | CreateGame");
 	return new Workspace();
 }